+++ src/main/java/com/petlife/shelter/dao/impl/ShelterBookingDAOImpl.java
@@ -114,6 +114,9 @@ public class ShelterBookingDAOImpl  implements Idao<ShelterBooking>{
 				if ("shelterId".equals(row.getKey())) {
 					predicates.add(builder.equal(root.get("shelterId"), row.getValue()));
 				}
+				if ("month".equals(row.getKey())) {
+					predicates.add(builder.greaterThanOrEqualTo(root.get("shelterBookingDate"), Date.valueOf(row.getValue())));
+				}
 					
 
 			}
